feat(owners): Additional owners are also set up as users (#193)
Additional owners are also set up as users, respecting the auth mode of
the container. If the user already exists in the database (e.g. because
the container used a bak or bacpac) and is disabled, it is enabled and
the password is set for NavUserPassword mode. If the user came from a
different domain (after the @ in the auth email), that is corrected. If
the user exists and is enabled, nothing is done